[usePageSize]refact/shared로 분리하여 Best와 OnSale에서 공용
