+++ sprint-mission/src/pages/ItemsPage/section/BestItems/BestItems.jsx
@@ -1,9 +1,18 @@
 import "./BestItems.css";
 import { ItemCard } from "../ui/ItemCard";
+import { SCREEN_SIZES } from "../../../../shared/hooks/useScreenSize";
+import { usePageSize } from "../../../../shared/hooks/usePageSize";
 import { useBestItems } from "./hooks/useBestItems";
 
+const SCREEN_SIZES_TO_PAGE_SIZE = {
+  [SCREEN_SIZES.MOBILE]: 1,
+  [SCREEN_SIZES.TABLET]: 2,
+  [SCREEN_SIZES.DESKTOP]: 4,
+};
+
 export function BestItems() {
-  const { productList } = useBestItems();
+  const { pageSize } = usePageSize(SCREEN_SIZES_TO_PAGE_SIZE); //반응형으로 pageSize 가져오기
+  const { productList } = useBestItems(pageSize);
 
   return (
     <section id="best-items">
